bodega_system/requirements.txt
asgiref==3.9.1
charset-normalizer==3.4.3
crispy-tailwind==1.0.3
Django==5.2.6
django-crispy-forms==2.4
django-htmx==1.24.1
django-simple-history==3.10.1
djangorestframework==3.16.1
pillow==11.3.0
psycopg2-binary==2.9.10
reportlab==4.4.3
sqlparse==0.5.3

# Testing dependencies
pytest==8.0.0
pytest-django==4.8.0
pytest-cov==4.1.0
factory-boy==3.3.0
freezegun==1.4.0
django-test-plus==2.2.3
